Split struct literal resolver tests
